Fix(memctrl): Fix the bug in asynchronization flag update when rdy low. The iDone and lsbDone flags cannot be reset when rdy is low. If so, those flags cannot be set again when rdy goes high.
That's fairly tricky. If unfortunately, the rdy goes low right after the
ifetch of memctrl is completed(call it clock #0), then the iDone flag is
perceived to be high by external in clock #1, just when the rdy is low.
So, the external will ignore this flag.
However, in old design, we will reset iDone flag to low in clock #1 (
perceived by external in clock #2).
Consequently, after the rdy signal goes high, the iDone signal is miss!
This is the bug. The same bug also exists in lsbDone flag.

I fix them both.